Validate user-provided job graphs
Previously, if a user attempted to assert a graph that did not fully
specify all dependencies such as

    {:bar [:quux]}

(here, where `:quux` is not specified in the graph), `bar`'s
dependencies would be incorrectly asserted and rejected by Datomic. Now,
graphs are validated before reaching Datomic, and the previous example
would cause an AssertionError. For this example to work, the user must
specify

    {:quux []
     :bar [:quux]}