Pass :pre-process return into process
Work was done in https://github.com/framed-data/overseer/pull/20 to
implement a 'harnessing' system, which allowed for finer-grained
instrumentation of the handler stage functions. However, this system was
implemented stemming from the need to transform input to a job prior to
execution, which is what a pre-processor is perfect for, of course.

The handler stage signatures are much more coherent when the default
pre-processor passes its input to :process, and :post-process returns
its result unmodified.