API: Fix identity function in harnessing
Previously when harnessing missing keys, the identity function was
supplied, returning the job it was passed. This is very useful for
uniformly harnessing a set of jobs which may or may not define the
particular lifecycle being harnessed.

However, identity (of arity 1) is only sufficient for 2/3 cases -
pre-process and process. post-process has arity 2 (job res) and we
need to use an appropriate identity function in that case.