Job reservation: Use standalone exception handler
Previously the job reservation function used the standard exception
handler, which logs errors to timbre and sends them to Sentry. This led
to lots of ineligible status exceptions in Sentry, which is a
non-actionable item that should be suppressed. The approach here is
changed to use a separate exception handler for job reservation, which
catches/logs ineligible starts and logs them before bailing out (no
Sentry).

By doing this, only 1 remaining caller of the default handler remained,
so we can just inline it with the standard job exception handler.