Remove all notion of a "system"
This removes all references to "systems", which is actually an
overloaded term in Overseer. The term was used to refer to both:

1) The overseer.system namespace, which confusingly combined a CLI
startup utility and an entry point to start a worker (Q: What's the
difference between a "system" and a "worker"? A: Nothing, really)

2) A map containing both :config and a :conn (Q: What's the difference
between a "system" map and a "config" map? A: Nothing, really)

The CLI startup code has been moved to `overseer.runner`, and the old
`system/start` function is removed in favor of the direct
`worker/start!` function (or the existing `overseer.api/start` alias, if
you like.) Additionally the signature has been changed to simply accept
a config and a set of job handlers - reason being that since `config`
already defines a Datomic URI, we can establish the connection ourselves
without requiring the user to do it and passing around a strange
"system" map.

Note that this also removes all of the old atom-based signalling
capabilities that were totally unused, and means that start is a
blocking call, i.e. it does not return a future that can be dereferenced.
This simplifies the code nicely, and can be safely removed as no
production workflow actually relies upon this behavior.

Future commits may re-add more sophisticated signalling/control
capabilities if deemed necessary.