(Experimental) Store job args in EDN
Previously, any user-specified arguments to jobs needed to already exist
as attributes in Datomic. This is undesirable for several reasons,
namely that it is inflexible, and frustratingly slow to iterate/change
as all possible attributes have to be enumerated and transacted into the
DB.  Furthermore it unnecessarily couples that aspect of the system to
Datomic, although backend indifference is not a design goal yet.

It is much better to store arbitrary job arguments in a single
system-defined EDN attribute, as arguments were just maps of primitive
datatypes anyways. This gives total flexibility to present and future
callers. The amount of code required to accomplish this in Overseer is
surprisingly small, although it will cause breaking changes in all
callers (both at job injection site and handlers) that will require work
to fix before rollout.

This is labeled as experimental because it is a low priority issue and
intended to serve for discussion purposes - for example, is modifying
the signature of `:process` handlers to be `[job args]` the right call?
All jobs have `:job/args` stored on them, but it seems lame to require
users to destruct/parse that themselves.