Manually specify source location
Prior to this commit, Leiningen would auto-infer the git source
information (which Clojars displays prominently).  For forks, this is
broken--it will think the canonical source comes from your fork, and
then display that fork in Clojars.  This wrong, so manually specify
the location of the upstream fork so Leiningen/Clojars knows to
display that, rather than any individual dev fork.