Modify lottery ratio: 3/2 => 4/1
Overseer uses lottery-style allocation to determine which eligible job
to run next. Previously, the preference ratio for unstarted over started
was 3/2. This modifies the ratio to 4/1 since we want to highly
prioritize unstarted work, and only towards the end of the queue tend
towards picking up anything that may died in-flight.