Explicitly require clj-http 1.1.2
This commit explicitly adds a dependency on at least clj-http 1.1.2.
Prior to this commit, this dependency was only implicitly coming in
from `raven-clj`, which was bringing in a very old version of
`clj-http` that was not compatible with Clojure 1.7.

This commit makes it so that Overseer, while not requiring the latest
clj-http or the latest Clojure 1.7, does require libraries that are at
least forward-compatible with Clojure 1.7.