Disable AOT compilation
I've confirmed that AOT is breaking dependency/Clojure 1.7
compatibility in our main project.  Because AOT is only used to enable
`overseer.runner`, which no one currently actually uses yet, this
commit disables AOT for this project as a library by default.
This unblocks our main project to upgrade to Clojure 1.7.

Note that this is specifically to serve the use case of Overseer as a
library in something else, which is the primary/only use case for
now--but you can still accomplish the goal of AOT compiling
`overseer.runner` using a trivial shim or by including the library
in your primary project and then designating `runner` as your main
and AOTing your primary project.

So, disabling AOT here enables much more flexibility for other
applications depending upon Overseer as a library, and one should
still be able to get the `runner` functionality if one chooses,
so theoretically this shouldn't be much of a tradeoff.