Errors: Catch/support java.util.concurrent.ExecutionException
If an exception is thrown within a future, it will die with a
j.u.c.ExecutionException 'wrapping' the original exception. This means
that if e.g. `overseer/abort` is used within a future, it will noisily
log to Sentry and follow the normal rules because the system is unable
to see that it's actually a controlled exception with ex-data. This
modifies the error handling to recognize j.u.c.ExecutionExceptions as we
want any thread to be able to control job execution, and supports the
normal rules.

There is ongoing discussion about whether or not exceptions are a
suitable means to 'communicate' known statuses with the Overseer
harness, vs a system oriented around return values and tagged data
structures or similar. Regardless of the conclusion of this discussion,
this is a hotfix of sorts for a current production situation.