Don't pass output of `pre-process` into `process`
The thinking here was the pre-process could be a place to modify context
and do dependency injection of sorts if desired; however it leads to
unnecessary API confusion ("even if you're just using it to set up
prerequisite state, make sure your pre-process function returns a job,
otherwise things will break mysteriously"), and harnesses are the
preferred mechanism to transform any job stage, not just pre-processors.

This is a breaking change for code that uses pre-processors to transform
jobs; however changing those harnesses to modify the process stage
instead should be trivial.