Serialize job arguments as EDN
This is a revision of https://github.com/framed-data/overseer/pull/42 ;
to recap:

"Previously, any user-specified arguments to jobs needed to already
exist as attributes in Datomic. This is undesirable for several reasons,
namely that it is inflexible, and frustratingly slow to iterate/change
as all possible attributes have to be enumerated and transacted into the
DB. Furthermore it unnecessarily couples that aspect of the system to
Datomic, although backend indifference is not a design goal yet."

In short, `:job/args` stores arbitrary as EDN and is transparently
se-/de- serialized. However, the design is slightly different compared
to #42, namely, the signature of process/post-process is *not* changed.
Instead, `(:job/args job)` is the standard interface to access the
parsed EDN. Given this design, it will be trivial to modify any existing
accessors (i.e. `->stream`) to check for the args attribute, or else
default to the old style as the change is being rolled out.