AOT-compile Overseer.runner
overseer.runner is used by the launcher script included with Overseer.
It used to be AOT-compiled, but this was disabled as AOT compilation
causes issues with applications that need to modify dependency versions
- see https://github.com/onyx-platform/onyx/issues/339. This re-enables
AOT-compilation for that namespace, but uses the same fix strategy of
requiring namespaces at runtime and resolving symbols which should
address the issue. This probably comes at some minor perf cost, but is
only run once at runtime.