API: Clarify naming of txn construction functions
Previously, the API functions for inserting new jobs used inconsistent
terms and notations: compare `->graph-txn` and `job-assertion`. This
unifies the names as `job-txn` and `graph-txns`.

The `txn` style naming was chosen as its more backend-independent, as
opposed to `-assertion` which is strongly coupled to Datomic.

This is technically a breaking change although only in symbol naming; no
semantics are changed.