Remove overseer.runner / overseer executable
The intent of overseer.runner (and the corresponding `overseer` binary)
was to give callers a way to start up overseer workers directly without
having to worry about defining worker namespaces/mains and pulling the Overseer
libraries into their own code. However, this execution model never
really panned out, and the implementation is inflexible in that it:

- Requires a YAML config file, doesn't integrate well with dynamic
  or environment config

- Requires `java` of the desired version to be on the PATH (not the end of the world,
  but not very flexible)

While overseer.runner may be convenient for certain situations, the benefit
is not enough to worry about maintaining it.