Remove option to disable worker heartbeats
Worker heartbeats used to be optional via the `:heartbeat :enabled`
config option; however the system was never really set up with an
appropriate alternative execution strategy if disabled. In theory we
could re-institute the old behavior of considering started jobs eligible
if heartbeating is disabled, but there is not a tremendous amount of
value in supporting this.