Make overseer.core/transact-graph idempotent on job-ids
In a distributed execution context it can be useful to use job-ids as
idempotence keys of sorts, and allow multiple graph transaction attempts
with only one succeeding. This makes store inserts idempontent on
job-ids; if multiple inserts with the same job-ids are attempted the
system will no-op and return the original graph without raising an
exception.